backend/requirements.txt
Django==2.2.18
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
